+++ src/App.tsx
@@ -7,7 +7,7 @@ import Update from './pages/proxy/Update';
 import Settings from './pages/proxy/Settings';
 import Delete from './pages/proxy/Delete';
 import Root from './pages/proxy/Root';
-import SettingsPeerList from './pages/proxy/SettingsPeerList';
+import SettingsAddConnections from './pages/proxy/SettingsAddConnections';
 
 function App() {
   return (
@@ -21,7 +21,7 @@ function App() {
           <Route path="/update" element={<Update />} />
           <Route path="/delete/:id" element={<Delete />} />
           <Route path="/settings" element={<Settings />} />
-          <Route path="/settings/peer-list" element={<SettingsPeerList />} />
+          <Route path="/settings/add-connections" element={<SettingsAddConnections />} />
           <Route path="/" element={<Root />} />
         </Routes>
       </AppProvider>