+++ src/pages/Dashboard/Settings/index.tsx
@@ -1,5 +1,5 @@
 import { useContext, useState } from 'react';
-import PeerList from './PeerList';
+import PeerList from './AddConnections';
 import Wallpaper from './Wallpaper';
 import ShutdownNode from './ShutdownNode';
 import { appContext } from '../../../AppContext';
@@ -9,22 +9,25 @@ import InScreenTitleBar from '../../../components/InScreenStatusBar';
 import DesktopConnect from './DesktopConnect';
 import UpdateMiniHub from './UpdateMiniHub';
 import { useNavigate } from 'react-router-dom';
+import ShareConnections from './ShareConnections';
 
 export function Settings() {
   const navigate = useNavigate();
-  const { setShowSettings, showSettings: display, setShowPeerList, showPeerList } = useContext(appContext);
+  const { setShowSettings, showSettings: display, showAddConnections, setShowAddConnections } = useContext(appContext);
   const [showShutdown, setShutdown] = useState(false);
   const [showWallpaper, setShowWallpaper] = useState(false);
   const [showDesktopConnect, setShowDesktopConnect] = useState(false);
   const [showBatteryOptimisation, setShowBatteryOptimisation] = useState(false);
   const [showUpdateMiniHub, setShowUpdateMinHub] = useState(false);
+  const [showShareConnections, setShowShareConnections] = useState(false);
 
   const dismiss = () => {
     navigate('/');
     setShowSettings(false);
     setShowUpdateMinHub(false);
-    setShowPeerList(false);
+    setShowAddConnections(false);
     setShowWallpaper(false);
+    setShowShareConnections(false);
     setShowDesktopConnect(false);
     setShowBatteryOptimisation(false);
   };
@@ -32,10 +35,11 @@ export function Settings() {
   return (
     <FullScreen display={display}>
       <ShutdownNode display={showShutdown} dismiss={() => setShutdown(false)} />
-      <PeerList display={showPeerList} dismiss={() => setShowPeerList(false)} />
       <Wallpaper display={showWallpaper} dismiss={() => setShowWallpaper(false)} />
-      <DesktopConnect display={showDesktopConnect} dismiss={() => setShowDesktopConnect(false)} />
+      <PeerList display={showAddConnections} dismiss={() => setShowAddConnections(false)} />
       <UpdateMiniHub display={showUpdateMiniHub} dismiss={() => setShowUpdateMinHub(false)} />
+      <DesktopConnect display={showDesktopConnect} dismiss={() => setShowDesktopConnect(false)} />
+      <ShareConnections display={showShareConnections} dismiss={() => setShowShareConnections(false)} />
       <BatteryOptimisation display={showBatteryOptimisation} dismiss={() => setShowBatteryOptimisation(false)} />
       <div className="flex flex-col h-full min-h-[700px]">
         <InScreenTitleBar onExit={dismiss} />
@@ -57,10 +61,24 @@ export function Settings() {
               </div>
             </div>
             <div
-              onClick={() => setShowPeerList(true)}
+              onClick={() => setShowShareConnections(true)}
               className="relative core-black-contrast-2 py-4 px-5 rounded cursor-pointer"
             >
-              Peer list
+              Share connections
+              <div className="absolute right-0 top-0 h-full px-5 flex items-center">
+                <svg width="8" height="12" viewBox="0 0 8 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+                  <path
+                    d="M7.04984 5.99995L1.37504 11.6501L0.500244 10.7501L5.24984 5.99995L0.500244 1.24975L1.40024 0.349747L7.04984 5.99995Z"
+                    fill="#F4F4F5"
+                  />
+                </svg>
+              </div>
+            </div>
+            <div
+              onClick={() => setShowAddConnections(true)}
+              className="relative core-black-contrast-2 py-4 px-5 rounded cursor-pointer"
+            >
+              Add connections
               <div className="absolute right-0 top-0 h-full px-5 flex items-center">
                 <svg width="8" height="12" viewBox="0 0 8 12" fill="none" xmlns="http://www.w3.org/2000/svg">
                   <path